requirements.txt
django==1.10
markdown==2.6.9
djangorestframework==3.6.4
celery==4.1.0
psycopg2==2.7.3.1
requests==2.18.4
pyyaml==3.12